feat: add filter option in datepicker and extract FilterBase from Filter
