+++ src/components/DatePicker/DatePicker.stories.mdx
@@ -90,6 +90,7 @@ Implementation of the regular Range DatePicker on different sizes
       <DatePicker
         inputProps={{ size: 'md', styleType: 'outlined' }}
         isRangePicker
+        isClearable={true}
         isDefaultNow={false}
         disableDates={select('disableDates', options, options[0])}
         dateFormatOverride={select('dateFormat', dateFormatOptions, dateFormatOptions[1])}
