+++ src/components/Filter/Filter.tsx
@@ -5,25 +5,12 @@ import React, { useMemo } from 'react';
 import { ChangeEvent } from 'utils/common';
 import { generateTestDataId } from 'utils/helpers';
 
-import { BASE_SHADE, pickTextColorFromSwatches } from '../../theme/palette';
-import Icon from '../Icon';
-import { AcceptedIconNames } from '../Icon/types';
 import ClickAwayListener from '../utils/ClickAwayListener';
+import FilterBase from './components/FilterBase';
 import Options from './components/Options/Options';
 import SearchInput from './components/SearchInput/SearchInput';
-import {
-  buttonSpanStyle,
-  buttonStyle,
-  buttonWrapperStyle,
-  childrenWrapperStyle,
-  dividedButtonStyle,
-  divider,
-  labelSpanStyle,
-  menuStyle,
-  valueSpanStyle,
-  wrapperStyle,
-} from './Filter.style';
-import { FilterOption, FilterType, Props, StyleType } from './types';
+import { menuStyle } from './Filter.style';
+import { FilterOption, Props } from './types';
 import { getTextColor } from './utils';
 import handleSearch from 'components/utils/handleSearch';
 
@@ -46,7 +33,7 @@ const Filter = React.forwardRef<HTMLButtonElement, Props>((props, ref) => {
     isVirtualized = false,
     onClear = () => {},
   } = props;
-  const [open, setOpen] = React.useState(false);
+  const [isOpen, setIsOpen] = React.useState(false);
   const [searchValue, setSearchValue] = React.useState('');
 
   const theme = useTheme();
@@ -56,15 +43,15 @@ const Filter = React.forwardRef<HTMLButtonElement, Props>((props, ref) => {
   const calculatedColor = calculateColorBetweenColorAndType('', buttonType);
 
   const iconColor = getTextColor({
-    open,
+    open: isOpen,
     theme,
     calculatedColor,
     hasSelectedValue,
   });
-  const iconName = open ? 'triangleUp' : 'triangleDown';
+  const iconName = isOpen ? 'triangleUp' : 'triangleDown';
 
   const handleSelect = (option: FilterOption) => {
-    setOpen(false);
+    setIsOpen(false);
     onSelect(option);
   };
 
@@ -95,7 +82,7 @@ const Filter = React.forwardRef<HTMLButtonElement, Props>((props, ref) => {
 
   const handleOpen = () => {
     setSearchValue('');
-    setOpen(!open);
+    setIsOpen(!isOpen);
   };
 
   const debouncedOnChange = React.useCallback(
@@ -114,24 +101,24 @@ const Filter = React.forwardRef<HTMLButtonElement, Props>((props, ref) => {
   }
 
   return (
-    <ClickAwayListener onClick={() => setOpen(false)}>
+    <ClickAwayListener onClick={() => setIsOpen(false)}>
       <FilterBase
         ref={ref}
-        dataTestId={generateTestDataId('filter', dataTestId)}
+        dataTestId={dataTestId}
         handleOpen={handleOpen}
         disabled={disabled}
         onClear={onClear}
         selectedItemLabel={selectedItem?.label ?? defaultValue.label}
-        open={open}
+        open={isOpen}
         hasSelectedValue={hasSelectedValue}
-        label={label}
+        label={label as string}
         iconName={iconName}
         iconColor={iconColor}
         filterType={filterType}
         buttonType={buttonType}
         styleType={styleType}
       >
-        {open && (
+        {isOpen && (
           <div css={menuStyle()} data-testid={generateTestDataId('filter-menu', dataTestId)}>
             {isSearchable && (
               <SearchInput
@@ -155,114 +142,6 @@ const Filter = React.forwardRef<HTMLButtonElement, Props>((props, ref) => {
         )}
       </FilterBase>
     </ClickAwayListener>
-  );
-});
-
-// eslint-disable-next-line react/display-name
-export const FilterBase = React.forwardRef<
-  HTMLButtonElement,
-  {
-    dataTestId: string;
-    disabled: boolean;
-    handleOpen: () => void;
-    onClear: () => void;
-    selectedItemLabel?: string;
-    open: boolean;
-    hasSelectedValue: boolean;
-    label?: string | undefined;
-    iconName: AcceptedIconNames;
-    iconColor: string;
-    iconSize?: number;
-    /** Defines the style type of the button */
-    styleType: StyleType;
-    /** Defines the filter type */
-    filterType?: FilterType;
-    buttonType?: 'primary' | 'secondary';
-  }
->((props, ref) => {
-  const {
-    dataTestId,
-    handleOpen,
-    disabled,
-    onClear,
-    selectedItemLabel,
-    open,
-    hasSelectedValue,
-    label,
-    iconName,
-    iconColor,
-    iconSize = 6,
-    buttonType = 'primary',
-    filterType = 'preset',
-    styleType,
-    children,
-  } = props;
-  const { calculateColorBetweenColorAndType } = useTypeColorToColorMatch();
-  const calculatedColor = calculateColorBetweenColorAndType('', buttonType);
-  const theme = useTheme();
-
-  const buttonStyleProps = {
-    calculatedColor,
-    buttonType,
-    disabled,
-    open,
-    styleType,
-    hasSelectedValue,
-    filterType,
-  };
-
-  const pickIconColor = () => {
-    if (open) {
-      return theme.utils.getColor('neutralWhite', 100);
-    }
-    if (hasSelectedValue) {
-      return theme.utils.getColor(calculatedColor.color, 550);
-    }
-
-    return theme.utils.getColor('darkGrey', 400);
-  };
-
-  return (
-    <div css={wrapperStyle()} data-testid={dataTestId}>
-      <button
-        ref={ref}
-        data-testid={generateTestDataId('filter', dataTestId)}
-        onClick={handleOpen}
-        disabled={disabled}
-        css={buttonWrapperStyle(buttonStyleProps)}
-      >
-        <div css={buttonStyle(buttonStyleProps)}>
-          <span css={buttonSpanStyle()}>
-            <span css={childrenWrapperStyle()}>
-              <span css={labelSpanStyle(open, hasSelectedValue)}>
-                {label && <div>{label} :</div>}
-                <span css={valueSpanStyle()}>{selectedItemLabel}</span>
-              </span>
-            </span>
-
-            <Icon name={iconName} color={iconColor} size={iconSize} />
-          </span>
-        </div>
-
-        {filterType === 'added' && (
-          <>
-            <span css={divider(buttonStyleProps)} />
-            <div css={dividedButtonStyle(buttonStyleProps)}>
-              <Icon
-                size={19}
-                name={'closeTag'}
-                color={pickIconColor()}
-                onClick={e => {
-                  e.stopPropagation();
-                  onClear();
-                }}
-              />
-            </div>
-          </>
-        )}
-      </button>
-      {children}
-    </div>
   );
 });
