+++ src/components/Filter/components/FilterBase/index.ts
@@ -1,0 +1,1 @@
+export { default } from './FilterBase';
